Add continous operation mode
Add a continous operation mode that keeps the
circulation pump active as long as the heat water
pump is active and we assume that hot water is still
being poured.

This mode supports faster warm water and helps in case
of water going in the wrong direction through the
circulation system/pump.